requirements.txt
django==5.1.1
djangorestframework==3.15.2
Markdown==3.7
django-timescaledb==0.2.13
djangorestframework_gis==1.1
psycopg[binary]==3.2.3
drf-yasg==1.21.7
